requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
bcrypt==4.0.1
python-dotenv==1.0.0
difflib==0.0.4
msgpack==1.0.5
redis==5.0.1
pytest==7.4.3 